Add conditions for user data dependent pages, must exist

+++ delete.php
@@ -5,8 +5,13 @@ require_once 'class' . DIRECTORY_SEPARATOR . 'User.php';
 
 session_start();
 
+if (!isset($_SESSION['user'])) {
+    header('Location: index.php');
+    die();
+}
 
-
-$task = new Task();
-
-$task->delete($_POST['delete']);
+if (isset($_POST['delete'])) {
+    $task = new Task();
+    
+    $task->delete($_POST['delete']);
+}
